COMMERCE-11854 Autogenerated code (blade gw buildService)

+++ modules/apps/commerce/commerce-price-list-api/src/main/java/com/liferay/commerce/price/list/model/CommerceTierPriceEntryModel.java
@@ -365,14 +365,14 @@ public interface CommerceTierPriceEntryModel
 	 *
 	 * @return the min quantity of this commerce tier price entry
 	 */
-	public int getMinQuantity();
+	public BigDecimal getMinQuantity();
 
 	/**
 	 * Sets the min quantity of this commerce tier price entry.
 	 *
 	 * @param minQuantity the min quantity of this commerce tier price entry
 	 */
-	public void setMinQuantity(int minQuantity);
+	public void setMinQuantity(BigDecimal minQuantity);
 
 	/**
 	 * Returns the display date of this commerce tier price entry.
