+++ modules/apps/commerce/commerce-price-list-api/src/main/java/com/liferay/commerce/price/list/service/persistence/CommerceTierPriceEntryPersistence.java
@@ -19,6 +19,8 @@ import com.liferay.commerce.price.list.model.CommerceTierPriceEntry;
 import com.liferay.portal.kernel.service.persistence.BasePersistence;
 import com.liferay.portal.kernel.service.persistence.change.tracking.CTPersistence;
 
+import java.math.BigDecimal;
+
 import java.util.Date;
 
 import org.osgi.annotation.versioning.ProviderType;
@@ -640,7 +642,7 @@ public interface CommerceTierPriceEntryPersistence
 	 * @throws NoSuchTierPriceEntryException if a matching commerce tier price entry could not be found
 	 */
 	public CommerceTierPriceEntry findByC_M(
-			long commercePriceEntryId, int minQuantity)
+			long commercePriceEntryId, BigDecimal minQuantity)
 		throws NoSuchTierPriceEntryException;
 
 	/**
@@ -651,7 +653,7 @@ public interface CommerceTierPriceEntryPersistence
 	 * @return the matching commerce tier price entry, or <code>null</code> if a matching commerce tier price entry could not be found
 	 */
 	public CommerceTierPriceEntry fetchByC_M(
-		long commercePriceEntryId, int minQuantity);
+		long commercePriceEntryId, BigDecimal minQuantity);
 
 	/**
 	 * Returns the commerce tier price entry where commercePriceEntryId = &#63; and minQuantity = &#63; or returns <code>null</code> if it could not be found, optionally using the finder cache.
@@ -662,7 +664,8 @@ public interface CommerceTierPriceEntryPersistence
 	 * @return the matching commerce tier price entry, or <code>null</code> if a matching commerce tier price entry could not be found
 	 */
 	public CommerceTierPriceEntry fetchByC_M(
-		long commercePriceEntryId, int minQuantity, boolean useFinderCache);
+		long commercePriceEntryId, BigDecimal minQuantity,
+		boolean useFinderCache);
 
 	/**
 	 * Removes the commerce tier price entry where commercePriceEntryId = &#63; and minQuantity = &#63; from the database.
@@ -672,7 +675,7 @@ public interface CommerceTierPriceEntryPersistence
 	 * @return the commerce tier price entry that was removed
 	 */
 	public CommerceTierPriceEntry removeByC_M(
-			long commercePriceEntryId, int minQuantity)
+			long commercePriceEntryId, BigDecimal minQuantity)
 		throws NoSuchTierPriceEntryException;
 
 	/**
@@ -682,7 +685,7 @@ public interface CommerceTierPriceEntryPersistence
 	 * @param minQuantity the min quantity
 	 * @return the number of matching commerce tier price entries
 	 */
-	public int countByC_M(long commercePriceEntryId, int minQuantity);
+	public int countByC_M(long commercePriceEntryId, BigDecimal minQuantity);
 
 	/**
 	 * Returns all the commerce tier price entries where commercePriceEntryId = &#63; and minQuantity &le; &#63;.
@@ -692,7 +695,7 @@ public interface CommerceTierPriceEntryPersistence
 	 * @return the matching commerce tier price entries
 	 */
 	public java.util.List<CommerceTierPriceEntry> findByC_LteM(
-		long commercePriceEntryId, int minQuantity);
+		long commercePriceEntryId, BigDecimal minQuantity);
 
 	/**
 	 * Returns a range of all the commerce tier price entries where commercePriceEntryId = &#63; and minQuantity &le; &#63;.
@@ -708,7 +711,7 @@ public interface CommerceTierPriceEntryPersistence
 	 * @return the range of matching commerce tier price entries
 	 */
 	public java.util.List<CommerceTierPriceEntry> findByC_LteM(
-		long commercePriceEntryId, int minQuantity, int start, int end);
+		long commercePriceEntryId, BigDecimal minQuantity, int start, int end);
 
 	/**
 	 * Returns an ordered range of all the commerce tier price entries where commercePriceEntryId = &#63; and minQuantity &le; &#63;.
@@ -725,7 +728,7 @@ public interface CommerceTierPriceEntryPersistence
 	 * @return the ordered range of matching commerce tier price entries
 	 */
 	public java.util.List<CommerceTierPriceEntry> findByC_LteM(
-		long commercePriceEntryId, int minQuantity, int start, int end,
+		long commercePriceEntryId, BigDecimal minQuantity, int start, int end,
 		com.liferay.portal.kernel.util.OrderByComparator<CommerceTierPriceEntry>
 			orderByComparator);
 
@@ -745,7 +748,7 @@ public interface CommerceTierPriceEntryPersistence
 	 * @return the ordered range of matching commerce tier price entries
 	 */
 	public java.util.List<CommerceTierPriceEntry> findByC_LteM(
-		long commercePriceEntryId, int minQuantity, int start, int end,
+		long commercePriceEntryId, BigDecimal minQuantity, int start, int end,
 		com.liferay.portal.kernel.util.OrderByComparator<CommerceTierPriceEntry>
 			orderByComparator,
 		boolean useFinderCache);
@@ -760,7 +763,7 @@ public interface CommerceTierPriceEntryPersistence
 	 * @throws NoSuchTierPriceEntryException if a matching commerce tier price entry could not be found
 	 */
 	public CommerceTierPriceEntry findByC_LteM_First(
-			long commercePriceEntryId, int minQuantity,
+			long commercePriceEntryId, BigDecimal minQuantity,
 			com.liferay.portal.kernel.util.OrderByComparator
 				<CommerceTierPriceEntry> orderByComparator)
 		throws NoSuchTierPriceEntryException;
@@ -774,7 +777,7 @@ public interface CommerceTierPriceEntryPersistence
 	 * @return the first matching commerce tier price entry, or <code>null</code> if a matching commerce tier price entry could not be found
 	 */
 	public CommerceTierPriceEntry fetchByC_LteM_First(
-		long commercePriceEntryId, int minQuantity,
+		long commercePriceEntryId, BigDecimal minQuantity,
 		com.liferay.portal.kernel.util.OrderByComparator<CommerceTierPriceEntry>
 			orderByComparator);
 
@@ -788,7 +791,7 @@ public interface CommerceTierPriceEntryPersistence
 	 * @throws NoSuchTierPriceEntryException if a matching commerce tier price entry could not be found
 	 */
 	public CommerceTierPriceEntry findByC_LteM_Last(
-			long commercePriceEntryId, int minQuantity,
+			long commercePriceEntryId, BigDecimal minQuantity,
 			com.liferay.portal.kernel.util.OrderByComparator
 				<CommerceTierPriceEntry> orderByComparator)
 		throws NoSuchTierPriceEntryException;
@@ -802,7 +805,7 @@ public interface CommerceTierPriceEntryPersistence
 	 * @return the last matching commerce tier price entry, or <code>null</code> if a matching commerce tier price entry could not be found
 	 */
 	public CommerceTierPriceEntry fetchByC_LteM_Last(
-		long commercePriceEntryId, int minQuantity,
+		long commercePriceEntryId, BigDecimal minQuantity,
 		com.liferay.portal.kernel.util.OrderByComparator<CommerceTierPriceEntry>
 			orderByComparator);
 
@@ -818,7 +821,7 @@ public interface CommerceTierPriceEntryPersistence
 	 */
 	public CommerceTierPriceEntry[] findByC_LteM_PrevAndNext(
 			long commerceTierPriceEntryId, long commercePriceEntryId,
-			int minQuantity,
+			BigDecimal minQuantity,
 			com.liferay.portal.kernel.util.OrderByComparator
 				<CommerceTierPriceEntry> orderByComparator)
 		throws NoSuchTierPriceEntryException;
@@ -829,7 +832,8 @@ public interface CommerceTierPriceEntryPersistence
 	 * @param commercePriceEntryId the commerce price entry ID
 	 * @param minQuantity the min quantity
 	 */
-	public void removeByC_LteM(long commercePriceEntryId, int minQuantity);
+	public void removeByC_LteM(
+		long commercePriceEntryId, BigDecimal minQuantity);
 
 	/**
 	 * Returns the number of commerce tier price entries where commercePriceEntryId = &#63; and minQuantity &le; &#63;.
@@ -838,7 +842,7 @@ public interface CommerceTierPriceEntryPersistence
 	 * @param minQuantity the min quantity
 	 * @return the number of matching commerce tier price entries
 	 */
-	public int countByC_LteM(long commercePriceEntryId, int minQuantity);
+	public int countByC_LteM(long commercePriceEntryId, BigDecimal minQuantity);
 
 	/**
 	 * Returns all the commerce tier price entries where displayDate &lt; &#63; and status = &#63;.
@@ -1159,7 +1163,7 @@ public interface CommerceTierPriceEntryPersistence
 	 * @return the matching commerce tier price entries
 	 */
 	public java.util.List<CommerceTierPriceEntry> findByC_LteM_S(
-		long commercePriceEntryId, int minQuantity, int status);
+		long commercePriceEntryId, BigDecimal minQuantity, int status);
 
 	/**
 	 * Returns a range of all the commerce tier price entries where commercePriceEntryId = &#63; and minQuantity &le; &#63; and status = &#63;.
@@ -1176,8 +1180,8 @@ public interface CommerceTierPriceEntryPersistence
 	 * @return the range of matching commerce tier price entries
 	 */
 	public java.util.List<CommerceTierPriceEntry> findByC_LteM_S(
-		long commercePriceEntryId, int minQuantity, int status, int start,
-		int end);
+		long commercePriceEntryId, BigDecimal minQuantity, int status,
+		int start, int end);
 
 	/**
 	 * Returns an ordered range of all the commerce tier price entries where commercePriceEntryId = &#63; and minQuantity &le; &#63; and status = &#63;.
@@ -1195,8 +1199,8 @@ public interface CommerceTierPriceEntryPersistence
 	 * @return the ordered range of matching commerce tier price entries
 	 */
 	public java.util.List<CommerceTierPriceEntry> findByC_LteM_S(
-		long commercePriceEntryId, int minQuantity, int status, int start,
-		int end,
+		long commercePriceEntryId, BigDecimal minQuantity, int status,
+		int start, int end,
 		com.liferay.portal.kernel.util.OrderByComparator<CommerceTierPriceEntry>
 			orderByComparator);
 
@@ -1217,8 +1221,8 @@ public interface CommerceTierPriceEntryPersistence
 	 * @return the ordered range of matching commerce tier price entries
 	 */
 	public java.util.List<CommerceTierPriceEntry> findByC_LteM_S(
-		long commercePriceEntryId, int minQuantity, int status, int start,
-		int end,
+		long commercePriceEntryId, BigDecimal minQuantity, int status,
+		int start, int end,
 		com.liferay.portal.kernel.util.OrderByComparator<CommerceTierPriceEntry>
 			orderByComparator,
 		boolean useFinderCache);
@@ -1234,7 +1238,7 @@ public interface CommerceTierPriceEntryPersistence
 	 * @throws NoSuchTierPriceEntryException if a matching commerce tier price entry could not be found
 	 */
 	public CommerceTierPriceEntry findByC_LteM_S_First(
-			long commercePriceEntryId, int minQuantity, int status,
+			long commercePriceEntryId, BigDecimal minQuantity, int status,
 			com.liferay.portal.kernel.util.OrderByComparator
 				<CommerceTierPriceEntry> orderByComparator)
 		throws NoSuchTierPriceEntryException;
@@ -1249,7 +1253,7 @@ public interface CommerceTierPriceEntryPersistence
 	 * @return the first matching commerce tier price entry, or <code>null</code> if a matching commerce tier price entry could not be found
 	 */
 	public CommerceTierPriceEntry fetchByC_LteM_S_First(
-		long commercePriceEntryId, int minQuantity, int status,
+		long commercePriceEntryId, BigDecimal minQuantity, int status,
 		com.liferay.portal.kernel.util.OrderByComparator<CommerceTierPriceEntry>
 			orderByComparator);
 
@@ -1264,7 +1268,7 @@ public interface CommerceTierPriceEntryPersistence
 	 * @throws NoSuchTierPriceEntryException if a matching commerce tier price entry could not be found
 	 */
 	public CommerceTierPriceEntry findByC_LteM_S_Last(
-			long commercePriceEntryId, int minQuantity, int status,
+			long commercePriceEntryId, BigDecimal minQuantity, int status,
 			com.liferay.portal.kernel.util.OrderByComparator
 				<CommerceTierPriceEntry> orderByComparator)
 		throws NoSuchTierPriceEntryException;
@@ -1279,7 +1283,7 @@ public interface CommerceTierPriceEntryPersistence
 	 * @return the last matching commerce tier price entry, or <code>null</code> if a matching commerce tier price entry could not be found
 	 */
 	public CommerceTierPriceEntry fetchByC_LteM_S_Last(
-		long commercePriceEntryId, int minQuantity, int status,
+		long commercePriceEntryId, BigDecimal minQuantity, int status,
 		com.liferay.portal.kernel.util.OrderByComparator<CommerceTierPriceEntry>
 			orderByComparator);
 
@@ -1296,7 +1300,7 @@ public interface CommerceTierPriceEntryPersistence
 	 */
 	public CommerceTierPriceEntry[] findByC_LteM_S_PrevAndNext(
 			long commerceTierPriceEntryId, long commercePriceEntryId,
-			int minQuantity, int status,
+			BigDecimal minQuantity, int status,
 			com.liferay.portal.kernel.util.OrderByComparator
 				<CommerceTierPriceEntry> orderByComparator)
 		throws NoSuchTierPriceEntryException;
@@ -1309,7 +1313,7 @@ public interface CommerceTierPriceEntryPersistence
 	 * @param status the status
 	 */
 	public void removeByC_LteM_S(
-		long commercePriceEntryId, int minQuantity, int status);
+		long commercePriceEntryId, BigDecimal minQuantity, int status);
 
 	/**
 	 * Returns the number of commerce tier price entries where commercePriceEntryId = &#63; and minQuantity &le; &#63; and status = &#63;.
@@ -1320,7 +1324,7 @@ public interface CommerceTierPriceEntryPersistence
 	 * @return the number of matching commerce tier price entries
 	 */
 	public int countByC_LteM_S(
-		long commercePriceEntryId, int minQuantity, int status);
+		long commercePriceEntryId, BigDecimal minQuantity, int status);
 
 	/**
 	 * Returns the commerce tier price entry where externalReferenceCode = &#63; and companyId = &#63; or throws a <code>NoSuchTierPriceEntryException</code> if it could not be found.
