+++ modules/apps/commerce/commerce-price-list-test/src/testIntegration/java/com/liferay/commerce/price/list/change/tracking/test/CommerceTierPriceEntryTableReferenceDefinitionTest.java
@@ -57,7 +57,7 @@ public class CommerceTierPriceEntryTableReferenceDefinitionTest
 	protected CTModel<?> addCTModel() throws Exception {
 		CommerceTierPriceEntryTestUtil.addCommerceTierPriceEntry(
 			_commercePriceEntry.getCommercePriceEntryId(),
-			RandomTestUtil.randomInt(), RandomTestUtil.randomDouble(),
+			RandomTestUtil.randomDouble(), RandomTestUtil.randomDouble(),
 			RandomTestUtil.randomDouble(), null);
 
 		return _commercePriceEntry;
